Renamed billing package to initialcredits.
