+++ api/src/main/webapp/WEB-INF/cron_base.yaml
@@ -16,7 +16,7 @@ cron:
   schedule: every 24 hours
   timezone: UTC
   target: api
-- description: Find and alert users that have exceeded their free tier billing usage using cloud task
+- description: Find and alert users that have exceeded their initial credit usage using cloud task
   url: /v1/cron/checkFreeTierBillingUsage
   schedule: every 30 minutes
   timezone: UTC
