+++ api/src/test/java/org/pmiops/workbench/initialcredits/InitialCreditsExpiryTaskMatchers.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.billing;
+package org.pmiops.workbench.initialcredits;
 
 import java.util.HashSet;
 import java.util.List;
